Move GridOptions to new types module
